+++ sdk/formrecognizer/Azure.AI.FormRecognizer/src/TrainingDocumentInfo.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System.Collections.Generic;
+using Azure.AI.FormRecognizer.Models;
 using Azure.Core;
 
 namespace Azure.AI.FormRecognizer.Training
@@ -11,6 +13,21 @@ namespace Azure.AI.FormRecognizer.Training
     [CodeGenModel("TrainingDocumentInfo")]
     public partial class TrainingDocumentInfo
     {
+        /// <summary> Initializes a new instance of TrainingDocumentInfo. </summary>
+        /// <param name="name"> Training document name. </param>
+        /// <param name="pageCount"> Total number of pages trained. </param>
+        /// <param name="errors"> List of errors. </param>
+        /// <param name="status"> Status of the training operation. </param>
+        /// <param name="modelId">The unique identifier of the model.</param>
+        internal TrainingDocumentInfo(string name, int pageCount, IReadOnlyList<FormRecognizerError> errors, TrainingStatus status, string modelId)
+        {
+            Name = name;
+            PageCount = pageCount;
+            Errors = errors;
+            Status = status;
+            ModelId = modelId;
+        }
+
         /// <summary>
         /// Training document name.
         /// </summary>
@@ -22,5 +39,10 @@ namespace Azure.AI.FormRecognizer.Training
         /// </summary>
         [CodeGenMember("pages")]
         public int PageCount { get; }
+
+        /// <summary>
+        /// The unique identifier of the model.
+        /// </summary>
+        public string ModelId { get; }
     }
 }