+++ sdk/formrecognizer/Azure.AI.FormRecognizer/CHANGELOG.md
@@ -1,7 +1,17 @@
 # Release History
 
 ## 3.1.0-beta.1 (Unreleased)
+
+### Breaking changes
 - It defaults to the latest supported API version, which currently is `2.1-preview.1`.
+
+### New Features
+- Added ability to create a composed model from the `FormTrainingClient` by calling method `StartCreateComposedModel`.
+- Added the properties `DisplayName` and `Properties` to types `CustomFormModel` and `CustomFormModelInfo`.
+- Added property `DisplayName` to `TrainingOptions` and new type `CreateComposedModelOptions`.
+- Added type `CustomFormModelProperties` that includes information like if a model is a composed model.
+- Added property `ModelId` to `CustomFormSubmodel` and `TrainingDocumentInfo`.
+- Added properties `ModelId` and `FormTypeConfidence` to `RecognizedForm`.
 - Added support to `StartRecognizeContent` to recognize selection marks such as check boxes and radio buttons.
 - Added support to train and recognize custom forms with selection marks such as check boxes and radio buttons. This functionality is only available in train with labels scenarios.